Move tests to plugin once more, where I think they belong
