+++ plugin/src/test/resources/io/jenkins/plugins/casc/core/JenkinsConfiguratorCloudSupportTest.yml
